Rename controllers folder by routers
