+++ lib/cartodb/routers/layergroup/analysis.js
