+++ lib/cartodb/routers/layergroup/attributes.js
