+++ lib/cartodb/routers/layergroup/dataview.js
