+++ lib/cartodb/routers/layergroup/middlewares/map-store-map-config-provider.js
