+++ lib/cartodb/routers/layergroup/preview.js
