+++ lib/cartodb/routers/layergroup/tile.js
