+++ lib/cartodb/routers/map/middlewares/augment-layergroup-data.js
