+++ lib/cartodb/routers/map/middlewares/check-json-content-type.js
