+++ lib/cartodb/routers/map/middlewares/increment-map-view-count.js
