+++ lib/cartodb/routers/map/middlewares/init-profiler.js
