+++ lib/cartodb/routers/map/middlewares/layer-stats.js
