+++ lib/cartodb/routers/map/middlewares/layergroup-id-header.js
