+++ lib/cartodb/routers/map/middlewares/layergroup-metadata.js
