+++ lib/cartodb/routers/map/middlewares/map-error.js
