+++ lib/cartodb/routers/map/named.js
