+++ lib/cartodb/routers/server-info.js
