+++ lib/cartodb/routers/template-router.js
