+++ lib/cartodb/routers/template/admin.js
