+++ lib/cartodb/routers/template/middlewares/named-map-provider.js
