+++ lib/cartodb/routers/template/preview.js
