+++ lib/cartodb/routers/template/tile.js
