+++ lib/cartodb/server.js
@@ -5,8 +5,8 @@ const { mapnik } = windshaft;
 
 const jsonReplacer = require('./utils/json-replacer');
 
-const ApiRouter = require('./controllers/api-router');
-const ServerInfoController = require('./controllers/server-info');
+const ApiRouter = require('./routers/api-router');
+const ServerInfoController = require('./routers/server-info');
 
 const StatsClient = require('./stats/client');
